add podman

+++ ubuntu_setup.sh
@@ -53,9 +53,19 @@ if [ -d ~/.oh-my-zsh ]; then
     sudo chsh -s "$(which zsh)"
 fi
 
+
+# Install podman if not present
+if ! dpkg -l | grep -q "^iiW\*podman"; then
+    echo "Installing podman"
+    sudo apt install -y podman
+else
+    echo "Podman is already installed. Skipping installation."
+fi
+
 # 
 # COCKPIT
 #
+
 # This requires switching to NetworkManager first
 netplan_file_path="/etc/netplan/01.netcfg.yaml"
 if [ ! -f "$netplan_file_path" ]; then
@@ -78,12 +88,14 @@ if [ ! -f "$netplan_file_path" ]; then
 else
     echo "Network manager already configured. No action taken."
 fi
-# Packages
-echo "Installing Cockpit packages"
-sudo apt install -y cockpit cockpit-pcp cockpit-machines cockpit-podman
-wget https://github.com/ocristopfer/cockpit-sensors/releases/latest/download/cockpit-sensors.deb
-sudo apt -f install -y ./cockpit-sensors.deb --fix-broken
-rm ./cockpit-sensors.deb
-wget https://github.com/45Drives/cockpit-benchmark/releases/download/v2.1.1/cockpit-benchmark_2.1.1-1focal_all.deb
-sudo apt -f install -y ./cockpit-benchmark_2.1.1-1focal_all.deb --fix-broken
-rm ./cockpit-benchmark_2.1.1-1focal_all.deb
+
+# Install cockpit if not present
+if ! dpkg -l | grep -q "^ii.*cockpit"; then
+    echo "Installing Cockpit packages"
+    sudo apt install -y cockpit cockpit-pcp cockpit-machines cockpit-podman
+    wget https://github.com/ocristopfer/cockpit-sensors/releases/latest/download/cockpit-sensors.deb
+    sudo apt -f install -y ./cockpit-sensors.deb --fix-broken
+    rm ./cockpit-sensors.deb
+else
+    echo "Cockpit is already installed. Skipping installation."
+fi
